+++ tests/Phinx/Db/Adapter/PdoAdapterTest.php
@@ -2,7 +2,6 @@
 
 namespace Test\Phinx\Db\Adapter;
 
-use Phinx\Db\Adapter\PdoAdapter;
 use PHPUnit\Framework\TestCase;
 
 class PdoAdapterTestPDOMock extends \PDO
@@ -58,8 +57,6 @@ class PdoAdapterTest extends TestCase
 
     public function testOptionsSetConnection()
     {
-        $this->assertNull($this->adapter->getConnection());
-
         $connection = new PdoAdapterTestPDOMock();
         $this->adapter->setOptions(['connection' => $connection]);